// src/indicador/domain/dtos/Indicador.dto.ts
export interface IndicadorDTO {
    quantidadeTfgs: number
    quantidadeAprovacoes: number
    quantidadeReprovacoes: number
    orientacoesRecusadas: number
    orientacoesAceitas: number
    entregasParciaisRealizadas: number
    entregasParciaisAprovadas: number
    entregasFinaisRealizadas: number
    orientacaoPendente: number
}